+++ src/Contracts/SchemaInterface.php
@@ -1,7 +1,9 @@
 <?php
 namespace DreamFactory\Core\Contracts;
 
-interface SchemaInterface extends CacheInterface, DbExtrasInterface
+use DreamFactory\Core\Database\Schema\TableSchema;
+
+interface SchemaInterface extends DbExtrasInterface
 {
     /**
      * Return an array of supported schema resource types.
@@ -31,22 +33,20 @@ interface SchemaInterface extends CacheInterface, DbExtrasInterface
      *
      * @param string $type    Resource type
      * @param string $schema  Schema name if any specific requested
-     * @param bool   $refresh Clear cache and retrieve anew?
      *
      * @return array
      */
-    public function getResourceNames($type, $schema = '', $refresh = false);
+    public function getResourceNames($type, $schema = '');
 
     /**
      * Return the metadata about a particular schema resource.
      *
      * @param string $type    Resource type
      * @param string $name    Resource name
-     * @param bool   $refresh Clear cache and retrieve anew?
      *
      * @return mixed
      */
-    public function getResource($type, $name, $refresh = false);
+    public function getResource($type, &$name);
 
     /**
      * @param string $type Resource type
@@ -55,6 +55,11 @@ interface SchemaInterface extends CacheInterface, DbExtrasInterface
      * @return mixed
      */
     public function dropResource($type, $name);
+
+    /**
+     * @return string|null
+     */
+    public function getDefaultSchema();
 
     /**
      * @return string|null
@@ -77,26 +82,89 @@ interface SchemaInterface extends CacheInterface, DbExtrasInterface
     public function isDefaultSchemaOnly();
 
     /**
-     * @param      $tables
-     * @param bool $allow_merge
-     * @param bool $allow_delete
-     * @param bool $rollback
+     * @param      $table
+     * @param array $options
      *
      * @return mixed
      */
-    public function updateSchema($tables, $allow_merge = false, $allow_delete = false, $rollback = false);
+    public function createTable($table, $options);
 
     /**
-     * Set the Caching interface.
+     * @param TableSchema $tableSchema
+     * @param array       $changes
      *
-     * @param  CacheInterface $cache
+     * @throws \Exception
      */
-    public function setCache($cache);
+    public function updateTable($tableSchema, $changes);
 
     /**
+     * @param      $table
+     *
      * @return mixed
      */
-    public function flushCache();
+    public function dropTable($table);
+
+    /**
+     * @param string       $table
+     * @param string|array $columns
+     *
+     * @return bool|int
+     */
+    public function dropColumns($table, $columns);
+
+    /**
+     * @param array $indexes
+     *
+     */
+    public function createFieldIndexes($indexes);
+
+    /**
+     * @param array $references
+     *
+     */
+    public function createFieldReferences($references);
+
+    /**
+     * @param string $table
+     * @param        $relationship
+     *
+     * @return bool|int
+     */
+    public function dropRelationship($table, $relationship);
+
+    public static function isUndiscoverableType($type);
+
+    public function getNamingSchema();
+
+    /**
+     * @param bool $unique
+     * @param bool $on_create_table
+     *
+     * @return bool
+     */
+    public function requiresCreateIndex($unique = false, $on_create_table = false);
+
+    /**
+     * @return bool
+     */
+    public function allowsSeparateForeignConstraint();
+
+    /**
+     * @param $table
+     * @param $column
+     *
+     * @return array
+     */
+    public function getPrimaryKeyCommands($table, $column);
+
+    /**
+     * @param string      $prefix
+     * @param string      $table
+     * @param string|null $column
+     *
+     * @return string
+     */
+    public function makeConstraintName($prefix, $table, $column = null);
 
     /**
      * Set the DB Extras interface.
@@ -104,11 +172,6 @@ interface SchemaInterface extends CacheInterface, DbExtrasInterface
      * @param  DbExtrasInterface $storage
      */
     public function setExtraStore($storage);
-
-    /**
-     * @return mixed
-     */
-    public function refresh();
 
     /**
      * @param string $name
@@ -151,4 +214,10 @@ interface SchemaInterface extends CacheInterface, DbExtrasInterface
      * @return string
      */
     public function getTimestampForSet();
+
+    /**
+     * @param string $name
+     * @return string
+     */
+    public function quoteIdentifier($name);
 }