+++ src/Models/BaseModel.php
@@ -690,7 +690,18 @@ class BaseModel extends Model implements CacheInterface
      */
     public function getTableSchema()
     {
-        return $this->getSchema()->getResource(DbResourceTypes::TYPE_TABLE, $this->table);
+        if (empty($result = $this->getFromCache('table:'.$this->table))) {
+            $schemaName = $this->getSchema()->getDefaultSchema();
+            $resourceName = $this->table;
+            $internalName = $schemaName . '.' . $resourceName;
+            $name = $resourceName;
+            $quotedName = $this->getSchema()->quoteIdentifier($schemaName) . '.' . $this->getSchema()->quoteIdentifier($resourceName);;
+            $settings = compact('schemaName', 'resourceName', 'name', 'internalName', 'quotedName');
+            $result = new TableSchema($settings);
+            $result = $this->getSchema()->getResource(DbResourceTypes::TYPE_TABLE, $result);
+            $this->addToCache('table:'.$this->table, $result, true);
+        }
+        return $result;
     }
 
     public function getSchema()